// src/app/models/property.ts
export class Property {
    id: string;
    name: string;
    purchasedate: any;
    initialcash: number;
    purchasevalue: number;
    currentcashbalance: number;
    currentvalue: number;
    budgetmanagement: number;
    budgetmortgage: number;
    budgetrates: number;
    budgetmaintenance: number;
    budgetinsurance: number;
    budgetutilities: number;
    budgetothercosts: number;
    budgetrentalincome: number;
    budgetrentalfrequency: string;
    budgetotherincome: number;
    budgetannualincome: number;
    actualmortgage: number;
    actualmanagement: number;
    actualrates: number;
    actualmaintenance: number;
    actualinsurance: number;
    actualutilities: number;
    actualothercosts: number;
    actualrentalincome: number;
    actualotherincome: number;
    actualannualincome: number;
    tenantname: string;
    tenantphone: string;
    tenantemail: string;
    contractdate: any;
    contractrenewal: any;
    taxable: string;
    forecast: boolean;
    notes: string;
}